Add parameters for enabling and disabling effects, and refactor effects to store a boolean for enabled effects rather than presence in a list
