+++ Source/audio/Effect.cpp
@@ -1,12 +1,12 @@
 #include "Effect.h"
 
-Effect::Effect(std::shared_ptr<EffectApplication> effectApplication, std::vector<EffectParameter> parameters) : effectApplication(effectApplication), parameters(parameters) {
+Effect::Effect(std::shared_ptr<EffectApplication> effectApplication, std::vector<EffectParameter> parameters) : effectApplication(effectApplication), parameters(parameters), enabled(parameters[0].name + " Enabled", parameters[0].id + "Enabled", true) {
 	smoothValues = std::vector<double>(parameters.size(), 0.0);
 }
 
 Effect::Effect(std::shared_ptr<EffectApplication> effectApplication, EffectParameter parameter) : Effect(effectApplication, std::vector<EffectParameter>{parameter}) {}
 
-Effect::Effect(std::function<Vector2(int, Vector2, const std::vector<double>&, double)> application, std::vector<EffectParameter> parameters) : application(application), parameters(parameters) {
+Effect::Effect(std::function<Vector2(int, Vector2, const std::vector<double>&, double)> application, std::vector<EffectParameter> parameters) : application(application), parameters(parameters), enabled(parameters[0].name + " Enabled", parameters[0].id + "Enabled", true) {
 	smoothValues = std::vector<double>(parameters.size(), 0.0);
 }
